Fixing pylint issues
